+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.3.0] - 2017-09-25
+
 ### Added
 - New request matchers for boolean composition:
   * `AndRequestMatcher`
@@ -36,6 +38,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 First release
 
-[Unreleased]: https://github.com/ajgarlag/psr15-router/compare/0.2.0...master
+[Unreleased]: https://github.com/ajgarlag/psr15-router/compare/0.3.0...master
+[0.3.0]: https://github.com/ajgarlag/psr15-router/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/ajgarlag/psr15-router/compare/0.1.1...0.2.0
 [0.1.1]: https://github.com/ajgarlag/psr15-router/compare/0.1.0...0.1.1